Minor changes to spelling and text display
Changed up a few things that i noticed when running this on a few deployments. The ask password was unformatted and echoing the color syntax, spelling, the while loop was only comparing itself to check GADMINPW. Also i noticed that if you're asking the user to set the grafana password at the beginning and then if they run your update grafana script then it will set it back to hunter2. Cheers!